+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - 2024-10-10
+
+### Added
+
+- Added hash of clientId, scopes and clientSecret to cacheKey of AccessToken
+
 ## [2.5.0] - 2024-08-13
 
 ### Added
@@ -19,7 +25,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Support for `preselected` provider selection under the Bank Transfer payment method
-- Support for `preselected` payment scheme under the preselected provider selection 
+- Support for `preselected` payment scheme under the preselected provider selection
 
 ## [2.3.0] - 2024-07-18